srorage: Finished remote database operations. 	 * Added deoxygen style comments to Remote and shared 	 storage interfaces as well as RemoteDatabaseOperations. 	 * Added implementation for remaining methods. 	 * Moved SharedDatabaseOperation method implementations 	 into LocalDatabaseOperations and deleted 	 SharedDatabaseOperations 	 * Change SharedDatabaseOperations and RemoteDatabaseOperations 	 to implements LocalDbOperations.

+++ app/src/main/java/storage/shapes/DatabaseOperations.java
@@ -43,7 +43,7 @@ public class DatabaseOperations  {
     }
 
     // return high score for user. returns integer high score
-    public int getHighScore(String username){
+    public long getHighScore(String username){
         return local.getHighScore(username);
 
     }
